ReEstimation fully implemented

+++ Project1/Structs.h
@@ -87,7 +87,7 @@ struct State
 
 	float * old_mean;
 	float * old_var;
-	int obs_count;
+	float obs_count;
 
 	int state_nr;
 	float g_const;
